Update plan generation logic for arbitrary data alignments
This adds information on alignment of the pdl->data regions
to the plan generation stage and provides the get_data_alignment()
routine in PDL::FFTW3 to return that value to the perl level.
The code appears to work correctly but the tests fail the
assumptions on whether or not the plan needs to be regenerated
are no longer correct for the fully general implementation.

Note, the current implementation does not support the use of
FFTW_MEASURE for plan generation.  Nor did the original code.